Moved camera_controls to src.utils.input_capture
