+++ src/gui/sections.py
@@ -8,6 +8,7 @@ from arcade.gui.widgets.layout import UIAnchorLayout, UIBoxLayout
 from arcade.gui.widgets.text import UILabel
 from pyglet.math import Vec2
 
+from src import config
 from src.engine.init_engine import eng
 from src.gui.buttons import CommandBarMixin
 from src.gui.combat_screen import CombatScreen
@@ -26,6 +27,7 @@ from src.gui.selection_texture_enums import SelectionCursor
 from src.gui.states import MissionCards
 from src.gui.ui_styles import ADVENTURE_STYLE
 from src.gui.window_data import WindowData
+from src.utils.camera_controls import CameraController
 from src.utils.pathing.grid_utils import Node
 
 
@@ -776,6 +778,7 @@ class CombatGridSection(arcade.Section):
     TILE_BASE_DIMS = (17, 16)
     SET_ENCOUNTER_HANDLER_ID = "set_encounter"
     SCALE_FACTOR = 0.2
+    GRID_ASPECT = (2.35, 1.1)
 
     def __init__(
         self,
@@ -829,16 +832,20 @@ class CombatGridSection(arcade.Section):
     @classmethod
     def init_path(cls, scale=1) -> arcade.SpriteList:
         selected_path_sprites = arcade.SpriteList()
-        start_sprite = arcade.Sprite(WindowData.indicators[SelectionCursor.GREEN.value], scale)
+        start_sprite = arcade.Sprite(
+            WindowData.indicators[SelectionCursor.GREEN.value], scale
+        )
         start_sprite.visible = False
         selected_path_sprites.append(start_sprite)
         for _ in range(1, 19):
             sprite_tex = WindowData.indicators[SelectionCursor.GOLD_EDGE.value]
-            sprite = arcade.Sprite(sprite_tex, scale=scale*WindowData.scale[1])
+            sprite = arcade.Sprite(sprite_tex, scale=scale * WindowData.scale[1])
             selected_path_sprites.append(sprite)
             sprite.visible = False
 
-        end_sprite = arcade.Sprite(WindowData.indicators[SelectionCursor.RED.value], scale)
+        end_sprite = arcade.Sprite(
+            WindowData.indicators[SelectionCursor.RED.value], scale
+        )
         selected_path_sprites.append(end_sprite)
         end_sprite.visible = False
 
@@ -894,10 +901,9 @@ class CombatGridSection(arcade.Section):
         grid_scale = 1
         sx, sy = WindowData.scale
         constant_scale = grid_scale * self.TILE_BASE_DIMS[0] * self.SCALE_FACTOR * 5
-        aspect = (2.35, 1.1)
         return Vec2(
-            (x - y) * sx * aspect[0],
-            (x + y) * sy * aspect[1],
+            (x - y) * sx * self.GRID_ASPECT[0],
+            (x + y) * sy * self.GRID_ASPECT[1],
         ) * constant_scale + Vec2(self.width / 2, 7 * self.height / 8)
 
     def grid_loc(self, v: Vec2) -> Node:
@@ -907,11 +913,11 @@ class CombatGridSection(arcade.Section):
         v2 = v - Vec2(self.width / 2, 7 * self.height / 8)
         sx, sy = self.scaling()
         grid_scale = 1
-        constant_scale = grid_scale * self.TILE_BASE_DIMS[0] * self.SCALE_FACTOR
+        constant_scale = grid_scale * self.TILE_BASE_DIMS[0] * self.SCALE_FACTOR * 5
         v3 = (
             Vec2(  # (x-y, x+y)
-                v2.x / (sx * 13),
-                v2.y / (sy * 5),
+                v2.x / (sx * self.GRID_ASPECT[0]),
+                v2.y / (sy * self.GRID_ASPECT[1]),
             )
             / constant_scale
         )
@@ -991,7 +997,14 @@ class CombatGridSection(arcade.Section):
         self.dudes_sprite_list.draw(pixelated=True)
 
         self.other_camera.use()
-        arcade.Text(repr(self.cam_controls), 10, WindowData.height - 20).draw()
+        if config.DEBUG:
+            arcade.Text(
+                repr(self.cam_controls),
+                10,
+                WindowData.height - 20,
+                multiline=True,
+                width=self.width,
+            ).draw()
 
         if eng.awaiting_input:
             self.combat_screen.draw_turn_prompt()
@@ -1104,98 +1117,3 @@ class CombatGridSection(arcade.Section):
         self.dudes_sprite_list.pop(
             self.dudes_sprite_list.index(dead_dude.owner.entity_sprite.sprite)
         )
-
-
-class CameraController:
-    _camera: arcade.Camera
-    _pan_speed: float
-    _zoom_increment: float
-    _pressed_keys: set[int]
-    _v_pan: set[int] = {
-        arcade.key.W,
-        arcade.key.S,
-    }
-    _h_pan: set[int] = {
-        arcade.key.D,
-        arcade.key.A,
-    }
-    _zoom: set[int] = {
-        arcade.key.Q,
-        arcade.key.E,
-    }
-
-    def __init__(self, camera: arcade.Camera, pan_speed=2.0, zoom_increment=0.002):
-        self._camera = camera
-        self._pan_speed = pan_speed
-        self._zoom_increment = zoom_increment
-        self._pressed_keys = set()
-
-    @property
-    def listened_keys(self) -> set[int]:
-        return self._v_pan | self._h_pan | self._zoom
-
-    def on_key_press(self, key: int) -> int | None:
-        include = {key} & self.listened_keys
-        self._pressed_keys |= include
-        if not include:
-            return key
-
-    def on_key_release(self, key) -> int | None:
-        include = {key} & self.listened_keys
-        self._pressed_keys -= include
-        if not include:
-            return key
-
-    @property
-    def zoom_factor(self) -> float:
-        z = 1
-        match [*(self._zoom & self._pressed_keys)]:
-            case [arcade.key.Q]:
-                z += self._zoom_increment
-            case [arcade.key.E]:
-                z -= self._zoom_increment
-            case _:
-                pass
-
-        return z
-
-    @property
-    def pan_vec(self) -> Vec2:
-        l, r, b, t = self._camera.projection
-        w, h = r - l, t - b
-
-        u, v = Vec2(w / 2, 0), Vec2(0, h / 2)
-        match [*(self._pressed_keys & self._h_pan)]:
-            # Right
-            case [arcade.key.D]:
-                x = 1
-            # Left
-            case [arcade.key.A]:
-                x = -1
-            # Both pressed or neither
-            case _:
-                x = 0
-
-        match [*(self._pressed_keys & self._v_pan)]:
-            # Up
-            case [arcade.key.W]:
-                y = 1
-            # Down
-            case [arcade.key.S]:
-                y = -1
-            # Both pressed or neither
-            case _:
-                y = 0
-
-        return (u * x + v * y).normalize()
-
-    def on_update(self):
-        l, r, b, t = self._camera.projection
-        w, h = r - l, t - b
-        center = self._camera.position + Vec2(1, 0)*w/2 + Vec2(0, 1)*h/2
-        self._camera.center(center + self.pan_vec*self._pan_speed, speed=1)
-        self._camera.zoom *= self.zoom_factor
-        self._camera.update()
-
-    def __repr__(self):
-        return "\n".join([f"{self._camera.position=}", f"{self._camera.zoom=}"])
